refactor: move github and gitlab to remote module
