+++ git-cliff-core/src/remote/mod.rs
@@ -1,0 +1,7 @@
+/// GitHub client.
+#[cfg(feature = "github")]
+pub mod github;
+
+/// GitLab client.
+#[cfg(feature = "gitlab")]
+pub mod gitlab;
